pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-facture-electronique-fr"
version = "0.2.0"
description = "MCP server exposing the AFNOR XP Z12-013 APIs for French electronic invoicing (Compatible Solution)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Christophe Méndez" }]
keywords = ["mcp", "facturation-electronique", "xp-z12-013", "ppf", "pa", "b2b"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp-einvoicing-core>=0.1.0",
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
]

[project.scripts]
mcp-facture-electronique-fr = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["clients", "tools"]
artifacts = ["server.py", "config.py"]

[tool.hatch.build.targets.sdist]
exclude = [
    "specs/",
    "tests/",
    ".env*",
    ".DS_Store",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
